+++ pages/api/song/current/[username].ts
@@ -32,6 +32,14 @@ async function handler(req: NextApiRequest, res: NextApiResponse<any>) {
 
 		const user = data[0];
 
+		if (user.preferences && user.preferences.current === false) {
+			return new BaseErrorResponse(
+				204,
+				'Current song is disabled in preferences',
+				{}
+			).handleResponse(req, res);
+		}
+
 		// Check if spotify_users is falsy
 		if (!user.spotify_users) {
 			return new BaseErrorResponse(